Semifunctional Firefox adaptation
I tweaked the query grabber to fix an issue where spaces in the title were replaced with random symbols,  so now all the sources that use search by name and not the IMDB id work as they should.

The tweak to use this on Firefox is to add a custom filter to uBlock, as I didn't find a way to adapt the declarativeNetRequest from Chrome API to the Firefox webRequest one.

The filter looks like this:

||letterboxd.com/csi/film/*/availability/$xhr,1p